Move ContactBuilder package
