+++ src/test/java/tagline/logic/LogicManagerTest.java
@@ -28,10 +28,10 @@ import tagline.model.ModelManager;
 import tagline.model.ReadOnlyAddressBook;
 import tagline.model.UserPrefs;
 import tagline.model.contact.Contact;
+import tagline.model.contact.ContactBuilder;
 import tagline.storage.JsonAddressBookStorage;
 import tagline.storage.JsonUserPrefsStorage;
 import tagline.storage.StorageManager;
-import tagline.testutil.ContactBuilder;
 
 public class LogicManagerTest {
     private static final IOException DUMMY_IO_EXCEPTION = new IOException("dummy exception");
