+++ src/test/java/tagline/logic/commands/CreateContactCommandIntegrationTest.java
@@ -12,7 +12,7 @@ import tagline.model.Model;
 import tagline.model.ModelManager;
 import tagline.model.UserPrefs;
 import tagline.model.contact.Contact;
-import tagline.testutil.ContactBuilder;
+import tagline.model.contact.ContactBuilder;
 
 /**
  * Contains integration tests (interaction with the Model) for {@code CreateContactCommand}.
