+++ src/test/java/tagline/testutil/TypicalContacts.java
@@ -15,6 +15,7 @@ import java.util.List;
 
 import tagline.model.AddressBook;
 import tagline.model.contact.Contact;
+import tagline.model.contact.ContactBuilder;
 
 /**
  * A utility class containing a list of {@code Contact} objects to be used in tests.
